moved resource_t to rman.h
