+++ include/bus.h
@@ -1,8 +1,7 @@
 #ifndef _SYS_BUS_H_
 #define _SYS_BUS_H_
 
-#include <common.h>
-#include <device.h>
+#include <rman.h>
 
 typedef struct resource resource_t;
 typedef struct bus_space bus_space_t;
@@ -41,34 +40,6 @@ struct bus_space {
 };
 
 #define BUS_SPACE_DECLARE(name) extern bus_space_t name[1]
-
-/* `resource` describes a range of addresses where a resource is mapped within
- * given bus space. A driver will use addresses from `r_start` to `r_end` and
- * `r_bus_space` routines to access hardware resource, so that the actual
- * driver code is not tied to way handled device is attached to the system. */
-
-#define RT_UNKNOWN 0
-#define RT_MEMORY 1
-#define RT_IOPORTS 2
-
-#define RF_NONE 0
-/* According to PCI specification prefetchable bit is CLEAR when memory mapped
- * resource contains locations with read side-effects or locations in which the
- * device does not tolerate write merging. */
-#define RF_PREFETCHABLE 1
-#define RF_SHARED 2 // TODO
-#define RF_ALLOCATED 4
-
-struct resource {
-  bus_space_t *r_bus_space; /* bus space accessor descriptor */
-  void *r_owner;            /* pointer to device that owns this resource */
-  intptr_t r_start;         /* first physical address of the resource */
-  intptr_t r_end; /* last (inclusive) physical address of the resource */
-  unsigned r_type;
-  unsigned r_flags;
-  int r_id; /* (optional) resource identifier */
-  LIST_ENTRY(resource) resources;
-};
 
 #define RESOURCE_DECLARE(name) extern resource_t name[1]
